Enumerated DataSizing::SizingFactorMode
